feat: refactor filter name
