+++ app/src/app/components/contents/common/Sidebar.tsx
@@ -11,7 +11,7 @@ export const Sidebar = ({ sidebarItems, children }: SidebarProps) => {
   const { pathname } = useLocation();
   return (
     <div className="flex flex-col space-y-8 lg:flex-row lg:space-x-10 lg:space-y-2">
-      <aside data-testid="sidebar" className="lg:w-1/5 pl-1 lg:pl-2">
+      <aside data-testid="sidebar" className="lg:w-1/6 pl-1 lg:pl-2">
         <nav className="flex space-x-2 lg:flex-col lg:space-x-0 lg:space-y-1">
           {sidebarItems.map((item, index) => (
             <NavLink
