+++ app/src/app/components/contents/manageAccess/ManageRoles.tsx
@@ -1,37 +1,18 @@
-import { SquarePlus, X } from "lucide-react";
+import { SquarePlus } from "lucide-react";
+import { useState } from "react";
 import { Button } from "../../Base/Button";
-import { Input } from "../../Base/Input";
 import { Skeleton } from "../../Base/Skeleton";
+import { useAuthConfig } from "../../providers/AuthConfigProvider";
 import { DataTable } from "../common/DataTable";
 import { ErrorComponent } from "../common/ErrorComponent";
+import { FilterByName } from "../common/FilterByName";
 import { useGetRolesWithRelationships } from "./hooks/useGetRolesWithRelationships";
 import { manageRolesColumns } from "./manageRolesColumns";
-import { useAuthConfig } from "../../providers/AuthConfigProvider";
-import { useCallback, useEffect, useRef, useState } from "react";
-import debounce from "lodash.debounce";
 
 export const ManageRoles = () => {
   const { roles, isLoading, error } = useGetRolesWithRelationships();
   const authConfig = useAuthConfig();
-  const [filterByName, setFilterByName] = useState("");
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const handleSetNameFilter = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setFilterByName(event.target.value);
-  };
-  const handleResetFilter = () => {
-    if (inputRef?.current) {
-      inputRef.current.value = "";
-    }
-    setFilterByName("");
-  };
-  const debouncedSetFilterByName = useCallback(debounce(handleSetNameFilter, 300), []);
-
-  useEffect(() => {
-    return () => {
-      debouncedSetFilterByName.cancel();
-    };
-  }, []);
+  const [filteredName, setFilterByName] = useState("");
 
   if (error) return <ErrorComponent message="Error fetching Roles" error={error} />;
 
@@ -56,20 +37,7 @@ export const ManageRoles = () => {
       </div>
       <div className="space-y-4 flex flex-col">
         <div className="flex justify-between">
-          <div className="flex space-x-4">
-            <Input
-              placeholder="Find Role by name..."
-              onChange={debouncedSetFilterByName}
-              className="h-8 sm:w-[450px] lg:w-[600px]"
-              ref={inputRef}
-            />
-            {filterByName && (
-              <Button variant="ghost" onClick={handleResetFilter} className="h-8 px-2 lg:px-3">
-                Reset
-                <X className="ml-2 h-4 w-4" />
-              </Button>
-            )}
-          </div>
+          <FilterByName filteredName={filteredName} setFilterByName={setFilterByName} />
           {authConfig?.enableBasicLogin && (
             <Button
               onClick={() => {
@@ -86,7 +54,7 @@ export const ManageRoles = () => {
         {roles && (
           <DataTable
             columns={manageRolesColumns}
-            data={roles.filter((role) => role.name.toLowerCase().includes(filterByName.toLowerCase()))}
+            data={roles.filter((role) => role.name.toLowerCase().includes(filteredName.toLowerCase()))}
             enablePagination
           />
         )}
